backend/requirements.txt
fastapi==0.128.0
uvicorn[standard]==0.40.0
sqlalchemy==2.0.46
pyodbc==5.3.0
pymysql==1.1.2
python-jose[cryptography]==3.5.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.22
pydantic==2.12.5
pydantic-settings==2.8.2
python-dotenv==1.0.1
